Added dry-run support.  Pre-refactoring cleanup.  Moved config file to /etc.

+++ sync.py
@@ -18,16 +18,15 @@ class BootSync:
 
     def __init__(self,api):
         self.api = api
+        self.verbose = True
 
     """
     Syncs the current bootconf configuration.  
-    Automatically runs the 'check_install'
+    Automatically runs the 'check' function first to eliminate likely failures.
     FUTURE: make dryrun work.
     """
-    def sync(self,dry_run=False):
-        if dry_run:
-            print "WARNING: dryrun hasn't been implemented yet.  Try not using dryrun at your own risk."
-            sys.exit(1)
+    def sync(self,dry_run=False,verbose=True):
+        self.dry_run = dry_run
         results = self.api.check()
         if results != []:
             self.api.last_error = m("run_check")
@@ -43,22 +42,38 @@ class BootSync:
             return False
         return True
 
+    """
+    Copy syslinux to the configured tftpboot directory
+    """
     def copy_pxelinux(self):
-        shutil.copy(self.api.config.pxelinux, os.path.join(self.api.config.tftpboot, "pxelinux.0"))
+        self.copy(self.api.config.pxelinux, os.path.join(self.api.config.tftpboot, "pxelinux.0"))
 
+
+    """
+    Delete any previously built pxelinux.cfg tree for individual systems.
+    This is better than trying to just add additional entries
+    as both MAC and IP settings could have been added and the MACs will
+    take precedence.  So we can't really trust human edits won't
+    conflict.
+    """
     def clean_pxelinux_tree(self):
-        shutil.rmtree(os.path.join(self.api.config.tftpboot, "pxelinux.cfg"), True) 
+        self.rmtree(os.path.join(self.api.config.tftpboot, "pxelinux.cfg"), True) 
 
+    """
+    A distro is a kernel and an initrd.  Copy all of them and error
+    out if any files are missing.  The conf file was correct if built
+    via the CLI or API, though it's possible files have been moved
+    since or perhaps they reference NFS directories that are no longer
+    mounted.
+    """
     def copy_distros(self):
         # copy is a 4-letter word but tftpboot runs chroot, thus it's required.
         images = os.path.join(self.api.config.tftpboot, "images")
-        shutil.rmtree(os.path.join(self.api.config.tftpboot, "images"), True)
-        os.mkdir(images)
+        self.rmtree(os.path.join(self.api.config.tftpboot, "images"), True)
+        self.mkdir(images)
         for d in self.api.get_distros().contents():
             kernel = self.api.utils.find_kernel(d.kernel) # full path
             initrd = self.api.utils.find_initrd(d.initrd) # full path
-            print "DEBUG: kernel = %s" % kernel
-            print "DEBUG: initrd = %s" % initrd
             if kernel is None:
                self.api.last_error = "Kernel for distro (%s) cannot be found and needs to be fixed: %s" % (d.name, d.kernel)
                raise "error"
@@ -67,9 +82,17 @@ class BootSync:
                raise "error"
             b_kernel = os.path.basename(kernel)
             b_initrd = os.path.basename(initrd)
-            shutil.copyfile(kernel, os.path.join(images, b_kernel))
-            shutil.copyfile(initrd, os.path.join(images, b_initrd))
+            self.copyfile(kernel, os.path.join(images, b_kernel))
+            self.copyfile(initrd, os.path.join(images, b_initrd))
 
+    """
+    Similar to what we do for distros, ensure all the kickstarts
+    in conf file are valid.  Since kickstarts are referenced by URL
+    (http or ftp), we do not have to copy them.  They are already
+    expected to be in the right place.  We can't check to see that the
+    URLs are right (or we don't, we could...) but we do check to see
+    that the files are at least still there.
+    """
     def validate_kickstarts(self):
         # ensure all referenced kickstarts exist
         # these are served by either NFS, Apache, or some ftpd, so we don't need to copy them
@@ -80,13 +103,18 @@ class BootSync:
               self.api.last_error = "Kickstart for group (%s) cannot be found and needs to be fixed: %s" % (g.name, g.kickstart)
               raise "error"
             
+    """
+    Now that kernels and initrds are copied and kickstarts are all valid,
+    build the pxelinux.cfg tree, which contains a directory for each 
+    configured IP or MAC address.
+    """
     def build_pxelinux_tree(self):
         # create pxelinux.cfg under tftpboot
         # and file for each MAC or IP (hex encoded 01-XX-XX-XX-XX-XX-XX)
         systems = self.api.get_systems()
         groups  = self.api.get_groups()
         distros = self.api.get_distros()
-        os.mkdir(os.path.join(self.api.config.tftpboot,"pxelinux.cfg"))
+        self.mkdir(os.path.join(self.api.config.tftpboot,"pxelinux.cfg"))
         for system in self.api.get_systems().contents():
             group = groups.find(system.group)
             if group is None:
@@ -100,6 +128,13 @@ class BootSync:
             filename = os.path.join(self.api.config.tftpboot, "pxelinux.cfg", filename)
             self.write_pxelinux_file(filename,system,group,distro)
 
+    """
+    The configuration file for each system pxelinux uses is either
+    a form of the MAC address of the hex version of the IP.  Not sure
+    about ipv6 (or if that works).  The system name in the config file
+    is either a system name, an IP, or the MAC, so figure it out, resolve
+    the host if needed, and return the pxelinux directory name.
+    """
     def get_pxelinux_filename(self,name_input):
         name = self.api.utils.find_system_identifier(name_input)
         if self.api.utils.is_ip(name):
@@ -110,24 +145,79 @@ class BootSync:
             self.api.last_error = "system name (%s) couldn't resolve and is not an IP or a MAC address." % name
             raise "error"
       
+    """
+    Write a configuration file for the pxelinux boot loader.
+    More system-specific configuration may come in later, if so
+    that would appear inside the system object in api.py
+    """
     def write_pxelinux_file(self,filename,system,group,distro):
         kernel_path = os.path.join("/images",os.path.basename(distro.kernel))
         initrd_path = os.path.join("/images",os.path.basename(distro.initrd))
         kickstart_path = self.api.config.kickstart_url + "/" + os.path.basename(group.kickstart)
-        file = open(filename,"w+")
-        file.write("default linux\n")
-        file.write("prompt 0\n")
-        file.write("timeout 1\n")
-        file.write("label linux\n")
-        file.write("   kernel %s\n" % kernel_path)
+        self.sync_log("writing: %s" % filename)
+        self.sync_log("---------------------------------")
+        if self.dry_run:
+            file = None
+        else:
+            file = open(filename,"w+")
+        self.tee(file,"default linux\n")
+        self.tee(file,"prompt 0\n")
+        self.tee(file,"timeout 1\n")
+        self.tee(file,"label linux\n")
+        self.tee(file,"   kernel %s\n" % kernel_path)
         # FIXME: leave off kickstart if no kickstart...
         # FIXME: allow specifying of other (system specific?) 
         #        parameters in bootconf.conf ???
-        file.write("   append devfs=nomount ramdisk_size=16438 lang= vga=788 ksdevice=eth0 initrd=%s ks=%s console=ttyS0,38400n8\n" % (initrd_path, kickstart_path))
-        file.close()
+        self.tee(file,"   append devfs=nomount ramdisk_size=16438 lang= vga=788 ksdevice=eth0 initrd=%s ks=%s console=ttyS0,38400n8\n" % (initrd_path, kickstart_path))
+        if not self.dry_run: 
+            file.close()
+        self.sync_log("--------------------------------")
 
+    """
+    For dry_run support, and logging...
+    """
+    def tee(self,file,text):
+        self.sync_log(text)
+        if not self.dry_run:
+            file.write(text)
+  
+    def copyfile(self,src,dst):
+       self.sync_log("copy %s to %s" % (src,dst))
+       if self.dry_run:
+           return True
+       return shutil.copyfile(src,dst)
+
+    def copy(self,src,dst):
+        self.sync_log("copy %s to %s" % (src,dst))
+        if self.dry_run:
+            return True
+        return shutil.copy(src,dst)
+
+    def rmtree(self,path,ignore):
+       self.sync_log("removing dir %s" % (path))
+       if self.dry_run:
+           return True
+       return shutil.rmtree(path,ignore)
+
+    def mkdir(self,path,mode=0777):
+       self.sync_log("creating dir %s" % (path))
+       if self.dry_run:
+           return True
+       return os.mkdir(path,mode)
+
+    """
+    Used to differentiate dry_run output from the real thing
+    automagically
+    """
+    def sync_log(self,message):
+        if self.verbose:
+            if self.dry_run:
+                print "dry_run | %s" % message
+            else:
+                print message
+            
+ 
     # FUTURE: would be nice to check if dhcpd and tftpd are running...
     # and whether kickstart url works (nfs, http, ftp)
     # at least those that work with open-uri
-    # possibly file permissions...
